Delegate exit processing to CompletionReporter.
In scenarios where we create multiple jasmine instances in a process, any global
listeners added need to be removed. Embedding jasmine in a development app is
on such scenario.

To ensure that the process on 'exit' listener is cleanly removed if and only
if the tests complete, we need to call removeListener in a jasmineDone() callback.
The CompletionReporter is the natural place to make the removal; then it makes
sense to attach the exit handler in a jasmineStarted() callback.

Further, to cleanly encapsulate exit handling, move the exit() calls to the
CompletionReporter. That simplifies the logic of exit handling as jasmine main.