Made the promise returned from Jasmine#execute usable
* An exitOnCompletion property to control whether Jasmine should make the
 Nnode process exit directly, rather than as a side effect of calling
 onComplete
* The promise returned from execute is resolved to the overall status
* onComplete is marked deprecated.

onComplete's behavior of installing a completion handler and also preventing
exit is clever, and actually what anyone using it probably wants, but it's
also non-obvious. The two-step process of setting exitOnCompletion = false
and then using the promise returned from execute() is not as slick, but it's
probably better for an audience that increasingly (for good reason) prefers
promise- or async/await-based async.