New API for exactly-once setup and teardown, especially in parallel mode
Top level `beforeAll`/`afterAll` are unavailable in parallel mode because
either running them exactly once in the parent process or once per worker
is going to be the wrong choice for a significant portion of existing uses.
The new `globalSetup` and `globalTeardown` config properties are run
exactly once regardless of whether the suite is run in parallel or not. They
are intended to support out-of-process initialization like starting an
external server.